:recycle: Split sign/verify between crypto/transactions
