+++ src/transactions/utils/prepareTransaction.js
@@ -12,20 +12,17 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-import cryptoModule from '../../crypto';
+import { signTransaction } from './signAndVerify';
 import getTransactionId from './getTransactionId';
 
 const secondSignTransaction = (transactionObject, secondSecret) =>
 	Object.assign({}, transactionObject, {
-		signSignature: cryptoModule.signTransaction(
-			transactionObject,
-			secondSecret,
-		),
+		signSignature: signTransaction(transactionObject, secondSecret),
 	});
 
 const prepareTransaction = (transaction, secret, secondSecret) => {
 	const singleSignedTransaction = Object.assign({}, transaction, {
-		signature: cryptoModule.signTransaction(transaction, secret),
+		signature: signTransaction(transaction, secret),
 	});
 
 	const signedTransaction =
